Script now reads from config file if values set, otherwise takes credentials from user input
Hopes, dreams, and security disclaimers added in README, along with a minor update on usage.
Aside from looking into if it would make sense to store the password encrypted,
next on the list would be to try to hook this script to local branch creation
and call the script with the name of the newly created branch.