Save updated branch list into file before script ends
All branches are listed and separated by a ", " delimiter (comma + space).
Currently useful so as to be able to `cat updated_branch_list | grep <branch_number>` while looking through active branches on github.

Perhaps this list can be read into an array in the script, along with a separate list read into another array using the output of `git branch -a -r`.
Any branches in the list that do not appear in `git branch -a -r` list should then be removed.